[MLOP-390] Create method to get feature sets' description (#188)
* Create reports module.

* Create metadata tests.

* Add mdutils in requirements

* Refactor method to sonar.

* Change variable names.

+++ tests/unit/butterfree/core/transform/test_feature_set.py
@@ -2,7 +2,7 @@ from unittest.mock import Mock
 
 import pytest
 from pyspark.sql import functions as F
-from pyspark.sql.types import FloatType, LongType, TimestampType
+from pyspark.sql.types import DoubleType, FloatType, LongType, TimestampType
 from tests.unit.butterfree.core.transform.conftest import (
     feature_add,
     feature_divide,
@@ -357,12 +357,12 @@ class TestFeatureSet:
             },
             {
                 "column_name": "feature1__stddev_pop_over_2_minutes_fixed_windows",
-                "type": FloatType(),
+                "type": DoubleType(),
                 "primary_key": False,
             },
             {
                 "column_name": "feature1__stddev_pop_over_15_minutes_fixed_windows",
-                "type": FloatType(),
+                "type": DoubleType(),
                 "primary_key": False,
             },
         ]
@@ -378,7 +378,7 @@ class TestFeatureSet:
                     transformation=SparkFunctionTransform(
                         functions=[
                             Function(F.avg, DataType.FLOAT),
-                            Function(F.stddev_pop, DataType.FLOAT),
+                            Function(F.stddev_pop, DataType.DOUBLE),
                         ]
                     ).with_window(
                         partition_by="id",
